norm lexer
