+++ Makefile
@@ -20,7 +20,7 @@ EXPANSION_SRCS = $(addprefix srcs/expansion/, \
 		expansion.c \
 		)
 LEXER_SRCS = $(addprefix srcs/lexer/, \
-		initialize_lexer.c lexer_utils.c lexer.c\
+		free_lexer_token.c initialize_lexer.c lexer_postprocess.c lexer_utils.c lexer.c\
 		)
 PARSER_SRCS = $(addprefix srcs/parser/, \
 		parser.c \
