+++ includes/lexer.h
@@ -6,7 +6,7 @@
 /*   By: takkatao <takkatao@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/23 16:12:06 by takkatao          #+#    #+#             */
-/*   Updated: 2022/04/23 16:12:06 by takkatao         ###   ########.fr       */
+/*   Updated: 2022/04/23 16:51:25 by takkatao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,4 +23,5 @@ bool			is_special_char(char c);
 void			set_token(	t_token	*t, int index, int token, char *word);
 void			set_state(t_lexer_manager *mgr, const char *str);
 void			free_lexer_token(t_token *token);
+void			lexer_postprocess(t_lexer_manager	*mgr);
 #endif
